remove sys_version for Python 3
since we're now committed to using Python3, I decided to go through and
clean up the code base to remove paths which depend on sys_version
checks for Python3
