+++ IPython/core/tests/test_magic_terminal.py
@@ -57,8 +57,6 @@ def check_cpaste(code, should_fail=False):
     finally:
         sys.stdin = stdin_save
 
-PY31 = sys.version_info[:2] == (3,1)
-
 def test_cpaste():
     """Test cpaste magic"""
 
@@ -77,13 +75,8 @@ def test_cpaste():
                       ],
 
              'fail': ["1 + runf()",
+                      "++ runf()",
              ]}
-    
-    # I don't know why this is failing specifically on Python 3.1. I've
-    # checked it manually interactively, but we don't care enough about 3.1
-    # to spend time fiddling with the tests, so we just skip it.
-    if not PY31:
-        tests['fail'].append("++ runf()")
 
     ip.user_ns['runf'] = runf
