+++ IPython/core/tests/test_ultratb.py
@@ -313,44 +313,43 @@ def r3o2():
 #----------------------------------------------------------------------------
 
 # module testing (minimal)
-if sys.version_info > (3,):
-    def test_handlers():
-        def spam(c, d_e):
-            (d, e) = d_e
-            x = c + d
-            y = c * d
-            foo(x, y)
+def test_handlers():
+    def spam(c, d_e):
+        (d, e) = d_e
+        x = c + d
+        y = c * d
+        foo(x, y)
 
-        def foo(a, b, bar=1):
-            eggs(a, b + bar)
+    def foo(a, b, bar=1):
+        eggs(a, b + bar)
 
-        def eggs(f, g, z=globals()):
-            h = f + g
-            i = f - g
-            return h / i
-        
-        buff = io.StringIO()
+    def eggs(f, g, z=globals()):
+        h = f + g
+        i = f - g
+        return h / i
 
-        buff.write('')
-        buff.write('*** Before ***')
-        try:
-            buff.write(spam(1, (2, 3)))
-        except:
-            traceback.print_exc(file=buff)
+    buff = io.StringIO()
 
-        handler = ColorTB(ostream=buff)
-        buff.write('*** ColorTB ***')
-        try:
-            buff.write(spam(1, (2, 3)))
-        except:
-            handler(*sys.exc_info())
-        buff.write('')
+    buff.write('')
+    buff.write('*** Before ***')
+    try:
+        buff.write(spam(1, (2, 3)))
+    except:
+        traceback.print_exc(file=buff)
 
-        handler = VerboseTB(ostream=buff)
-        buff.write('*** VerboseTB ***')
-        try:
-            buff.write(spam(1, (2, 3)))
-        except:
-            handler(*sys.exc_info())
-        buff.write('')
+    handler = ColorTB(ostream=buff)
+    buff.write('*** ColorTB ***')
+    try:
+        buff.write(spam(1, (2, 3)))
+    except:
+        handler(*sys.exc_info())
+    buff.write('')
+
+    handler = VerboseTB(ostream=buff)
+    buff.write('*** VerboseTB ***')
+    try:
+        buff.write(spam(1, (2, 3)))
+    except:
+        handler(*sys.exc_info())
+    buff.write('')
