+++ IPython/testing/skipdoctest.py
@@ -1,26 +1,13 @@
-"""Decorators marks that a doctest should be skipped, for both python 2 and 3.
+"""Decorators marks that a doctest should be skipped.
 
 The IPython.testing.decorators module triggers various extra imports, including
 numpy and sympy if they're present. Since this decorator is used in core parts
 of IPython, it's in a separate module so that running IPython doesn't trigger
 those imports."""
 
-#-----------------------------------------------------------------------------
-#  Copyright (C) 2009-2011  The IPython Development Team
-#
-#  Distributed under the terms of the BSD License.  The full license is in
-#  the file COPYING, distributed as part of this software.
-#-----------------------------------------------------------------------------
+# Copyright (C) IPython Development Team
+# Distributed under the terms of the Modified BSD License.
 
-#-----------------------------------------------------------------------------
-# Imports
-#-----------------------------------------------------------------------------
-
-import sys
-
-#-----------------------------------------------------------------------------
-# Decorators
-#-----------------------------------------------------------------------------
 
 def skip_doctest(f):
     """Decorator - mark a function or method for skipping its doctest.
