+++ tools/github_stats.py
@@ -114,9 +114,6 @@ if __name__ == "__main__":
     print("DEPRECATE: backport_pr.py is deprecated and is is now recommended"
           "to install `ghpro` from PyPI.", file=sys.stderr)
 
-    # deal with unicode
-    if sys.version_info < (3,):
-        sys.stdout = codecs.getwriter('utf8')(sys.stdout)
     
     # Whether to add reST urls for all issues in printout.
     show_urls = True
